requirements.txt
streamlit
paho-mqtt
python-dotenv
web3